washguard-prototype/requirements.txt
streamlit
pandas
matplotlib
wordcloud
transformers
torch
twilio
python-dotenv
